+++ Plugin.sln
@@ -266,6 +266,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WeaponPlus", "src\WeaponPlu
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dummy", "src\Dummy\Dummy.csproj", "{411A8F53-FEE2-4DC0-9965-7303FCBCB2F4}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AIChatPlugin", "src\AIChatPlugin\AIChatPlugin.csproj", "{A07C06F0-BE16-483B-BF8C-A3FEC49ADAC4}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -1306,6 +1308,14 @@ Global
 		{411A8F53-FEE2-4DC0-9965-7303FCBCB2F4}.Release|Any CPU.Build.0 = Release|Any CPU
 		{411A8F53-FEE2-4DC0-9965-7303FCBCB2F4}.Release|x64.ActiveCfg = Release|Any CPU
 		{411A8F53-FEE2-4DC0-9965-7303FCBCB2F4}.Release|x64.Build.0 = Release|Any CPU
+		{A07C06F0-BE16-483B-BF8C-A3FEC49ADAC4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A07C06F0-BE16-483B-BF8C-A3FEC49ADAC4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A07C06F0-BE16-483B-BF8C-A3FEC49ADAC4}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{A07C06F0-BE16-483B-BF8C-A3FEC49ADAC4}.Debug|x64.Build.0 = Debug|Any CPU
+		{A07C06F0-BE16-483B-BF8C-A3FEC49ADAC4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A07C06F0-BE16-483B-BF8C-A3FEC49ADAC4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A07C06F0-BE16-483B-BF8C-A3FEC49ADAC4}.Release|x64.ActiveCfg = Release|Any CPU
+		{A07C06F0-BE16-483B-BF8C-A3FEC49ADAC4}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
