+++ src/AIChatPlugin/AIChatPlugin.csproj
@@ -1,5 +1,3 @@
 <Project Sdk="Microsoft.NET.Sdk">
-
-<lmport Project="..\..\template.targets" />
-
+    <Import Project="..\..\template.targets" />
 </Project>
