+++ src/AIChatPlugin/AiType.cs
@@ -1,0 +1,14 @@
+﻿namespace AIChatPlugin;
+
+public enum AiType
+{
+    /// <summary>
+    /// 默认
+    /// </summary>
+    Default,
+
+    /// <summary>
+    /// 速度模式
+    /// </summary>
+    Fast
+}
